add upgrade to test
