+++ e2e/internal/upgrade/BUILD.bazel
@@ -10,9 +10,19 @@ go_library(
     importpath = "github.com/edgelesssys/constellation/v2/e2e/internal/upgrade",
     visibility = ["//e2e:__subpackages__"],
     deps = [
+        "//internal/api/attestationconfigapi",
+        "//internal/config",
         "//internal/constants",
+        "//internal/file",
+        "//internal/imagefetcher",
         "//internal/logger",
         "//internal/semver",
+        "//internal/versions",
+        "@com_github_spf13_afero//:afero",
+        "@com_github_stretchr_testify//require",
+        "@io_bazel_rules_go//go/runfiles:go_default_library",
+        "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
+        "@io_k8s_client_go//kubernetes",
         "@sh_helm_helm_v3//pkg/action",
         "@sh_helm_helm_v3//pkg/cli",
     ],
@@ -35,16 +45,9 @@ go_test(
     tags = ["manual"],
     deps = [
         "//e2e/internal/kubectl",
-        "//internal/api/attestationconfigapi",
-        "//internal/config",
         "//internal/constants",
-        "//internal/file",
-        "//internal/imagefetcher",
-        "//internal/semver",
         "//internal/versions",
-        "@com_github_spf13_afero//:afero",
         "@com_github_stretchr_testify//require",
-        "@io_bazel_rules_go//go/runfiles:go_default_library",
         "@io_k8s_api//core/v1:core",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
         "@io_k8s_client_go//kubernetes",
