+++ e2e/provider-upgrade/BUILD.bazel
@@ -1,0 +1,15 @@
+load("//bazel/go:go_test.bzl", "go_test")
+
+go_test(
+    name = "provider-upgrade_test",
+    srcs = ["upgrade_test.go"],
+    # keep
+    gotags = ["e2e"],
+    tags = ["manual"],
+    deps = [
+        "//e2e/internal/kubectl",
+        "//e2e/internal/upgrade",
+        "//internal/constants",
+        "@com_github_stretchr_testify//require",
+    ],
+)
